:bug: 削除できるのは、自分の言葉だけ

+++ app/controllers/users/personal_words_controller.rb
@@ -20,6 +20,8 @@ class Users::PersonalWordsController < UsersController
   end
 
   def destroy(id)
+    return unless current_user.personal_words.pluck(:id).include?(id)
+
     personal_word = PersonalWord.find_by(id: id)
     personal_word.destroy!
     redirect_to users_personal_words_path
